Adds KingMoveFinder and tests
The SlideMoveFinder was modified to accept a maximum range.  A king
has a maximum range of 1 and all other pieces which use the SlideMoveFinder
have unlimited range.  This allows KingMoveFinder to reuse the same logic
for not moving off the board, not moving into friendly pieces etc.

Finding castling moves is not yet implemented.

No piece can make a move which places its own king in check.  We will
filter these moves once all of the potential moves are found.